+++ src/App.js
@@ -7,7 +7,7 @@ import { Route, Switch } from 'react-router-dom';
 import './App.css';
 
 import Header from './components/Header/Header';
-import PicturesCollection from './components/PicturesCollection/PicturesCollection';
+import PostsManager from './components/PostsManager/PostsManager';
 import NewPost from './components/NewPost/NewPost';
 import { fetchPosts } from './store/actions';
 
@@ -20,7 +20,7 @@ function App(props) {
     <div className="App">
       <Header />
       <Switch>
-        <Route path="/" exact component={PicturesCollection} />
+        <Route path="/" exact component={PostsManager} />
         <Route path="/new-post" exact component={NewPost} />
       </Switch>
     </div>
