+++ src/components/PostsManager/PostsManager.module.css
@@ -1,4 +1,4 @@
-.PicturesCollection {
+.PostsManager {
     padding: 20px;
     display: flex;
     justify-content: flex-start;
